chore: refactor the inner logic
